+++ docs/RELEASE-NOTES.md
@@ -1,5 +1,18 @@
 # Release Notes
 
+## 4.0.0 - ?
+
+Breaking changes:
+- Fixed `TruncateResponse` option so inline with API default (now true by default).
+- Fixed `IncludeUnverified` option so inline with API default (now true by default).
+
+New features:
+- (None)
+
+Bug fixes / internal changes:
+- Removed package dependency `ByteDev.Common`.
+- Added package dependency `ByteDev.ResourceIdentifier`.
+
 ## 3.0.0 - 24 July 2020
 
 Breaking changes: